+++ redwood/api/src/services/cachedProfiles/cachedProfiles.ts
@@ -1,4 +1,4 @@
-import {CachedProfile as PrismaCachedProfile, StatusEnum} from '@prisma/client'
+import {CachedProfile as PrismaCachedProfile} from '@prisma/client'
 import {db} from 'src/lib/db'
 import {importProfile} from 'src/tasks/syncStarknetState'
 import {
@@ -7,6 +7,7 @@ import {
   QuerycachedProfilesArgs,
 } from 'types/graphql'
 import {ContractCache} from '../contractCache/contractCache'
+import {currentStatus, isVerified} from './helpers'
 
 export const cachedProfiles = async ({
   first,
@@ -56,81 +57,6 @@ export const cachedProfileByEthereumAddress = async ({
     where: {ethereumAddress},
   })
 
-// Keep in sync with `StatusEnum` in `starknet/contracts/profile.cairo`
-const STATUS_ENUMS: {[enumVal: number]: StatusEnum} = {
-  0: StatusEnum.NOT_CHALLENGED,
-  1: StatusEnum.CHALLENGED,
-  2: StatusEnum.ADJUDICATION_ROUND_COMPLETED,
-  3: StatusEnum.APPEALED,
-  4: StatusEnum.APPEAL_OPPORTUNITY_EXPIRED,
-  5: StatusEnum.SUPER_ADJUDICATION_ROUND_COMPLETED,
-  6: StatusEnum.SETTLED,
-}
-
-export const parseChallengeStatus = (status: number): StatusEnum =>
-  STATUS_ENUMS[status]
-
-// Keep in sync with profile.cairo#_get_current_status
-
-export const currentStatus = (
-  profile: Pick<
-    PrismaCachedProfile,
-    | 'lastRecordedStatus'
-    | 'challengeTimestamp'
-    | 'adjudicationTimestamp'
-    | 'appealTimestamp'
-  >,
-  now = new Date()
-): StatusEnum => {
-  if (profile.lastRecordedStatus === StatusEnum.NOT_CHALLENGED) {
-    return StatusEnum.NOT_CHALLENGED
-  } else if (profile.lastRecordedStatus === StatusEnum.CHALLENGED) {
-    const timePassed =
-      now.getTime() - (profile.challengeTimestamp?.getTime() ?? 0)
-    const hasAppealOpportunityExpired =
-      ContractCache.adjudicationTimeWindow + ContractCache.appealTimeWindow <
-      timePassed
-    if (hasAppealOpportunityExpired)
-      return StatusEnum.APPEAL_OPPORTUNITY_EXPIRED
-
-    const hasAdjudicationOpportunityExpired =
-      ContractCache.adjudicationTimeWindow < timePassed
-    if (hasAdjudicationOpportunityExpired)
-      return StatusEnum.ADJUDICATION_ROUND_COMPLETED
-
-    return StatusEnum.CHALLENGED
-  } else if (
-    profile.lastRecordedStatus === StatusEnum.ADJUDICATION_ROUND_COMPLETED
-  ) {
-    const timePassed =
-      now.getTime() - (profile.adjudicationTimestamp?.getTime() ?? 0)
-
-    const hasAppealOpportunityExpired =
-      ContractCache.appealTimeWindow < timePassed
-    if (hasAppealOpportunityExpired)
-      return StatusEnum.APPEAL_OPPORTUNITY_EXPIRED
-
-    return StatusEnum.ADJUDICATION_ROUND_COMPLETED
-  } else if (profile.lastRecordedStatus === StatusEnum.APPEALED) {
-    const timePassed = now.getTime() - (profile.appealTimestamp?.getTime() ?? 0)
-
-    const hasSuperAdjudicationOpportunityExpired =
-      ContractCache.superAdjudicationTimeWindow < timePassed
-    if (hasSuperAdjudicationOpportunityExpired)
-      return StatusEnum.SUPER_ADJUDICATION_ROUND_COMPLETED
-
-    return StatusEnum.APPEALED
-  } else if (
-    profile.lastRecordedStatus === StatusEnum.SUPER_ADJUDICATION_ROUND_COMPLETED
-  ) {
-    return StatusEnum.SUPER_ADJUDICATION_ROUND_COMPLETED
-  } else if (profile.lastRecordedStatus === StatusEnum.SETTLED) {
-    return StatusEnum.SETTLED
-  } else {
-    throw new Error(`Unknown profile status: ${profile.lastRecordedStatus}`)
-  }
-}
-
 // Keep in sync with profile.cairo#get_is_in_provisional_time_window
 const isInProvisionalTimeWindow = (
   profile: Pick<PrismaCachedProfile, 'submissionTimestamp'>,
@@ -138,42 +64,6 @@ const isInProvisionalTimeWindow = (
 ): boolean => {
   const timePassed = now.getTime() - profile.submissionTimestamp.getTime()
   return timePassed < ContractCache.provisionalTimeWindow
-}
-
-// Keep in sync with profile.cairo#_get_is_verified
-export const isVerified = (
-  profile: Pick<
-    PrismaCachedProfile,
-    | 'lastRecordedStatus'
-    | 'notarized'
-    | 'challengeTimestamp'
-    | 'superAdjudicationTimestamp'
-    | 'didAdjudicatorVerifyProfile'
-    | 'didSuperAdjudicatorOverturnAdjudicator'
-    | 'adjudicationTimestamp'
-    | 'appealTimestamp'
-    | 'submissionTimestamp'
-  >,
-  now = new Date()
-): boolean => {
-  const status = currentStatus(profile, now)
-  if (status == StatusEnum.NOT_CHALLENGED) {
-    const isProvisional = isInProvisionalTimeWindow(profile, now)
-    return profile.notarized || !isProvisional
-  } else if (status == StatusEnum.CHALLENGED) {
-    const isPresumedInnocent =
-      ContractCache.provisionalTimeWindow <
-      (profile.challengeTimestamp?.getTime() ?? 0) -
-        profile.submissionTimestamp.getTime()
-
-    return isPresumedInnocent
-  } else if (profile.superAdjudicationTimestamp != null) {
-    return profile.didSuperAdjudicatorOverturnAdjudicator
-      ? !profile.didAdjudicatorVerifyProfile
-      : profile.didAdjudicatorVerifyProfile
-  } else if (profile.adjudicationTimestamp != null) {
-    return profile.didAdjudicatorVerifyProfile
-  } else return false
 }
 
 export const CachedProfile = {
