+++ redwood/scripts/syncStarknetState.ts
@@ -2,5 +2,5 @@ import syncStarknetState from '$api/src/tasks/syncStarknetState'
 
 export default async ({args}) => {
   console.log('Syncing StarkNet state...')
-  await syncStarknetState(!!args.onlyNew)
+  await syncStarknetState({onlyNewProfiles: !!args.onlyNew})
 }
